scripts/requirements.txt
# ─── AI / LLM ────────────────────────────────────────────────
anthropic==0.86.0
openai==2.30.0

# ─── Voice ───────────────────────────────────────────────────
elevenlabs==2.40.0

# ─── Video & Audio ───────────────────────────────────────────
moviepy==2.2.1
faster-whisper==1.2.1
imageio-ffmpeg==0.6.0
av==17.0.0

# ─── B-Roll Generators ───────────────────────────────────────
# browser_visit: headless browser screenshot + FFmpeg scroll animation
# IMPORTANT: after pip install, run: playwright install chromium
playwright>=1.50.0
Pillow==11.3.0
Pygments==2.19.2

# image_montage: image fetching + Ken Burns slideshow
httpx==0.28.1
beautifulsoup4==4.14.3
lxml==6.0.2
feedparser==6.0.12       # Google News RSS thumbnails fallback

# audio duration measurement (optional — pipeline degrades gracefully without it)
# mutagen>=1.45.0        # not currently in venv; uncomment to improve duration accuracy

# ─── HTTP & Networking ───────────────────────────────────────
requests==2.33.0
aiohttp==3.13.3
websockets==16.0

# ─── Social & Platform APIs ──────────────────────────────────
python-telegram-bot==22.7
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0

# ─── GPU Cloud ───────────────────────────────────────────────
runpod==1.8.2

# ─── CLI & Utilities ─────────────────────────────────────────
click==8.3.1
rich==14.3.3
python-dotenv==1.2.2
python-dateutil==2.9.0.post0
schedule==1.2.2
pytrends==4.9.2           # niche analysis notebook

# ─── Development & Testing ───────────────────────────────────
pytest>=7.4.0
